Improve ability to run RM_Call commands as specific users/ACLs
Currently, RM_Call really runs commands as the NULL/root user.

While it can do ACL checks, it only does them at RM_Call time on the command issued itself.  So if one RM_Call's an eval, it will check the permission to run EVAL, but will not do any further checks during execution of the script, as the scipt will be run as the NULL user.

This adds an RM_SetContextModulUser that enables module witers to attach a RedisModuleUser to the context and if set use that ModuleUser for execution and ACL validation.

In order to allow module writers to both chose to run as the NULL/Root user or unset which user it should run as, we also provide a NullUser object to set when one desires the Null/Root user vs resetting with the NULL value.

Currently, the Redis Module API only allows one to set single ACL "ops" at a time via RM_SetModuleUserACL.

In order to allow module writers to more easily manipulate ACL strings, we provide a RM_SetModuleUserACLString.

In order to allow module writers to serialize out the ACL string we provide RM_GetModuleUserACLString.  This string representation is cached within Redis, invalidated on ACL change and only generated on demand.