allow modules blocking on keys to reply outside of reply_callback
the current code sets up blocking commands always as needing to be resent, but removes the flag when the reply_callback returns successfully.

However, if one is only using the reply callback as a signal that the keys are ready (i.e. using threads and thread safe contexts), it wont reset.

In practice, we don't need to set the flag if the blocker is a module, as we always would reset it before redis would try to reissue the commands.